+++ src/Parser/XMLDaoParser.php
@@ -80,7 +80,7 @@ class XMLDaoParser
     /**
      * @var string name of the class that should the dao factory should inherits from.
      */
-    private $parentFactoryClass = '\jDaoFactoryBase'; //'\Jelix\Dao\AbstractDaoFactory';
+    private $parentFactoryClass = '\Jelix\Dao\AbstractDaoFactory';
 
     /**
      * selector of the imported dao.